requirements.txt
astroid==2.4.2
attrs==20.3.0
certifi==2020.12.5
click==7.1.2
click-plugins==1.1.1
cligj==0.7.1
cycler==0.10.0
decorator==4.4.2
Fiona==1.8.18
geographiclib==1.50
geopandas==0.8.1
geopy==2.1.0
IMDbPY==2020.9.25
isort==5.7.0
kiwisolver==1.3.1
lazy-object-proxy==1.4.3
lxml==4.6.2
Mako==1.1.4
MarkupSafe==1.1.1
matplotlib==3.3.3
mccabe==0.6.1
munch==2.5.0
networkx==2.5
numpy==1.19.5
pandas==1.1.5
pep8==1.7.1
Pillow==8.1.0
pkg-resources==0.0.0
plotly==4.14.1
pycodestyle==2.6.0
pylint==2.6.0
pyparsing==2.4.7
pyproj==3.0.0.post1
python-dateutil==2.8.1
pytz==2020.5
retrying==1.3.3
scipy==1.5.4
seaborn==0.11.1
Shapely==1.7.1
six==1.15.0
SQLAlchemy==1.3.22
toml==0.10.2
typed-ast==1.4.2
wrapt==1.12.1
